Fixed bug in class PGaussianMean
When d was less than 1 (but not equal to 0) the neighbour values' weights
trumped the mid value weight (set to 1).
To fix this the algorithm first builds a vector of distances, then finds the
maximum distance and normalize all the other distances according to the
formula:

    dw = dmin / d

where is the current value of the distance vector, dmin is the minimum non-zero
distance and dw is the distance weight. The new weights are then computed
according to the formula:

    w = wg * dw

Where wg are the gaussian weights computed against the indices.