Added a new class member to Bio_Float class
The default_mean member will hold a Mean object that will be applyed to the
data read by read() method unless the users sets the mean parameter.
If default_mean is None and no mean object has been passed to read() then no
filtering will be perfomed on the data.