IMPORTANT BUG FIX : introducing ID() as method of profile objects
A profile cannot be indentified by its only name
In case of oceanographic cruise a name (e.g. Meteor) can be associated
several points in the same "model time" (e.g. a month)

Then, the ID() method provides all the informations to be unique.
ID() is used in Matchup_manager._dump_punti_for_aveScan()

set() is not more used in Matchup_manager.writefiles_for_profiling() because it shuffles the list
for debug it is better to have a (orderered) list