Dataextractor is now able to find by itself whether dimvar is 2 or 3D.
This solves this problem: build_layer_maps.py uses MapBuilder.plot_maps_data() which calls dataextractor without passing dimvar argument. Then, there is a conflict when actual dimvar is 2, while dataextractor sets as default 3.

Then, now dataextractor does not need dimvar as argument, but returns this info in output as De.dims.

It is a WIP, dimvar remains at the moment as argument, it will be removed in next commits.

Mapbuilder takes in account dataextractor.dims in to manage also 2D fields.