Bug fix in instrument/check.oxygen_check() Remove unuseful printouts
M.getMatchups2() is no more under try block. Issues have to be clear, printing ‘p.ID()  not found in PROFILATORE’ is not a good idea.

For debug reasons, we have to run each variable separately: for that reason we remove ivar, and we define a class ‘variable’ to define every feature
- name
- extra
- check object

inside the 'variable' object