Improved DataExtractor class
This commit solves a bug in the DataExtractor class and improves
its performance in some corner cases.

Before this commit, the DataExtractor class failed if a variable
had a dimension 1 on an axis with index i if there was an axis
with index j of dimension > 1 such that j < i. This should be
fixed with this commit.

Moreover, it cleans a little bit the code making some statements
more readable.

Finally, it also improves the performance of the classes when
we read the first layer of a 3D variable (i.e., when we submit
the argument dimvar=2). Before, it read all the variable and
then it discarded all the layers different from 0. Now the
routine reads only the required data.