Added a new version of the multirun plotter
This commit adds a new version of the profiles_plotter for the
multiruns. This new version adds a few new features: the most
noticeable is the possibility to add filters to data, for example
computing the moving mean.

Moreover, it is now also possible to define a different number of
levels (so, for example, 3 or 5 instead of 4) and it is also
possible to plot the average between multiple levels instead of a
fixed one.

Finally, it also introduces the possibility to compare more than
4 different plots together