Improved performances of the bathymetric regions
This commit introduces two new algorithms to check if a point
is inside a bathymetric region. The first algorithm works for
the regular masks and uses the routine `search_closest_sorted`
implemented in the utils.py file.

The other one uses a KDTree to speed up the search of the closest
center cell for a point defined by its coordinates and it is
useful for the non-regular masks.