DataExtractor object does not raise anymore error if mask depth and data depth are inconsistent.
When we work with reduced mask, data field will be sliced in order to fit mask.
When we work with reduced data,  data field will be increased with zeros in order to fit mask.

In these cases a warning is raised and can be silenced putting verbose=False when object is constructed.