Allowing bit.sea to read correctly the CADEAU mask
Some masks contain the values for dz only on the water cells; on
the other hand, bit.sea used to read the values of dz on the first
cell on the top-left corner.
This commit changes this behaviour: now bit.sea reads the values
for dz in the deepest column available in the mask.